Updated generated requirements
Generated using updated Makefile and freeze script from previous commit

Note that base.txt dependency on South is clarified to exclude (!=) 0.8
(which only works in Django 1.5) but allow later releases (but not
future South 2.0 release, which will be incompatible).  This allows use
of South 0.8.1 on Python >= 2.6; `pip install -r requirements/py25.txt`
before `pip install -r requirements/base.txt` is required for Python 2.5
(as is documented in the installation guide).

http://south.readthedocs.org/en/latest/releasenotes/0.8.1.html
http://south.readthedocs.org/en/latest/releasenotes/0.8.html